fix: 스케줄러 트랜잭션 필요 로직 별도 분리
